Move checknix script to test/xcompat
